requirements.txt
instructor
python-dotenv
rich